% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding-wkb.R
\name{as_wkbarrow}
\alias{as_wkbarrow}
\alias{as_wkbarrow.sf}
\alias{as_wkbarrow.sfc}
\title{Coerce object into an Arrow WKB array}
\usage{
as_wkbarrow(.x, ...)

\method{as_wkbarrow}{sf}(.x, ...)

\method{as_wkbarrow}{sfc}(.x, ...)
}
\arguments{
\item{.x}{\code{sf} or \code{sfc} object}

\item{...}{Unused}
}
\value{
An \link[arrow:Table-class]{arrow::Table} or \link[arrow:array-class]{arrow::Array}
}
\description{
Coerce object into an Arrow WKB array
}
