% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding-geoarrow.R
\name{.as_geoarrow_helper}
\alias{.as_geoarrow_helper}
\title{Recursive helper for geoarrow coercion.}
\usage{
.as_geoarrow_helper(x)
}
\description{
If \code{x} is a \code{list}, then this function recursively
applies itself to its elements. If \code{x} is a matrix,
we assume its a coordinate matrix based on \code{sf}'s
geometry structure, and convert it to a \code{data.frame}.
Otherwise, if \code{x} is neither of those, then we simply
return it unclassed.
}
\keyword{internal}
