% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georeed-utils.R
\name{geometry_columns}
\alias{geometry_columns}
\title{Get the names of geometry columns in an \code{sf} object}
\usage{
geometry_columns(.data)
}
\arguments{
\item{.data}{\code{sf}}
}
\value{
\code{character(N)}
Names of all geometry columns in the \code{sf} object.
}
\description{
Get the names of geometry columns in an \code{sf} object
}
\keyword{internal}
