% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georeed-parquet-metadata.R
\name{geoparquet_file_metadata}
\alias{geoparquet_file_metadata}
\title{Create GeoParquet file metadata}
\usage{
geoparquet_file_metadata(version, primary_column, columns)
}
\arguments{
\item{version}{\emph{\code{character(1)}}\cr
The version identifier for the GeoParquet specification.}

\item{primary_column}{\emph{\code{character(1)}}\cr
The name of the "primary" geometry column.
In cases where a GeoParquet file contains multiple geometry columns,
the primary geometry may be used by default in geospatial operations.}

\item{columns}{\emph{\code{list}}\cr
Metadata about geometry columns.
Each key is the name of a geometry column in the table.}
}
\value{
\code{list}
}
\description{
Create GeoParquet file metadata
}
\details{
Specification from https://geoparquet.org/releases/v1.0.0/
}
\seealso{
geoparquet_column_metadata
}
\keyword{internal}
