% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoparquet-metadata.R
\name{geoparquet_metadata}
\alias{geoparquet_metadata}
\title{Create GeoParquet metadata}
\usage{
geoparquet_metadata(.data, ..., encoding = "WKB", covering = FALSE)
}
\arguments{
\item{.data}{\emph{\code{sf}}\cr
Data to base metadata from.}

\item{...}{Named arguments for extra metadata added
to the parquet file, see details.}

\item{encoding}{\emph{\code{character(1)}}\cr
Type of encoding that the geometry will be in.
Must be either \emph{WKB} or one of:
\itemize{
\item \emph{point}
\item \emph{linestring}
\item \emph{polygon}
\item \emph{multipoint}
\item \emph{multilinestring}
\item \emph{multipolygon}
}}

\item{covering}{\emph{\code{logical(1)}}\cr
Experimental metadata option to include
geometry envelopes as a new parquet column.
If \code{TRUE}, assumes the bbox column was generated by
this package.}
}
\value{
\code{list} of the named arguments and the values coerced to JSON strings.
}
\description{
Create GeoParquet metadata
}
\details{
Extra named arguments passed to this function will be passed
to a \code{list} call for the metadata added to a GeoParquet file.
To specify additional custom metadata, each named argument
must be convertible to JSON.
}
\keyword{internal}
