% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoparquet-parquet.R
\name{read_geoparquet}
\alias{read_geoparquet}
\title{Read a GeoParquet file}
\usage{
read_geoparquet(
  file,
  col_select = NULL,
  as_data_frame = TRUE,
  props = arrow::ParquetArrowReaderProperties$create(),
  mmap = TRUE,
  ...
)
}
\arguments{
\item{file}{A character file name or URI, \code{raw} vector, an Arrow input stream,
or a \code{FileSystem} with path (\code{SubTreeFileSystem}).
If a file name or URI, an Arrow \link[arrow]{InputStream} will be opened and
closed when finished. If an input stream is provided, it will be left
open.}

\item{col_select}{A character vector of column names to keep, as in the
"select" argument to \code{data.table::fread()}, or a
\link[tidyselect:eval_select]{tidy selection specification}
of columns, as used in \code{dplyr::select()}.}

\item{as_data_frame}{Should the function return a \code{tibble} (default) or
an Arrow \link[arrow]{Table}?}

\item{props}{\link[arrow]{ParquetArrowReaderProperties}}

\item{mmap}{Use TRUE to use memory mapping where possible}

\item{...}{
  Arguments passed on to \code{\link[arrow:read_parquet]{arrow::read_parquet}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
An \code{sf} table if \code{as_data_frame} is \code{TRUE} (the default), or an
\link[arrow:Table-class]{arrow::Table} otherwise.
}
\description{
'\href{https://geoparquet.org/}{GeoParquet}' is a geospatial columnar
format backed by the '\href{https://parquet.apache.org/}{Parquet}' file format.
}
